app/src/requirements/production.txt
-r base.txt
django-celery==3.2.2
django-redis-cache==1.7.1
uWSGI==2.0.17.1
redis==2.10.6
